refactor(tasks): isolate SQLite connection kernels
Keep task and flow row codecs and SQL on supplied connections while preserving global database acquisition, cache/close ownership, transaction admission, and existing API return contracts. Move compound callers directly to the kernels and keep terminal-status classification with pure record types. Add cold-import and real SQLite compound rollback proof; no worker activation or schema change.
