+++ src/agents/subagents/completion/subagent-completion-admission.store.ts
@@ -24,7 +24,7 @@ import {
   bindTaskRecord,
   readTaskRecord,
   upsertTaskRunRowInDatabase,
-} from "../../../tasks/task-registry.store.sqlite.js";
+} from "../../../tasks/task-registry.store.kernel.js";
 import type { TaskRecord } from "../../../tasks/task-registry.types.js";
 import { ensureDeliveryState } from "../registry/subagent-delivery-state.js";
 import { resolveFinalizedSubagentTaskState } from "../registry/subagent-registry-completion.js";
