+++ src/tasks/task-registry-publication.ts
@@ -1,4 +1,3 @@
-import { isTerminalTaskStatus } from "./task-executor-policy.js";
 import { clearTaskActivity, flushTaskActivity } from "./task-registry-activity.js";
 import {
   cloneTaskRecord,
@@ -17,7 +16,7 @@ import {
   rebuildRunIdIndex,
   tasks,
 } from "./task-registry-state.js";
-import type { TaskRecord } from "./task-registry.types.js";
+import { isTerminalTaskStatus, type TaskRecord } from "./task-registry.types.js";
 
 /** Publishes a record already committed by a cross-owner shared-state transaction. */
 export function publishTaskRecordAfterAtomicStore(
